Add model and data module

+++ modules/feature_engineering.py
@@ -93,9 +93,9 @@ class FeatureEngineeringV1:
 		X_train = torch.FloatTensor(X_train)
 		X_val = torch.FloatTensor(X_val)
 		X_test = torch.FloatTensor(X_test)
-		y_train = torch.LongTensor(y_train)
-		y_val = torch.LongTensor(y_val)
-		y_test = torch.LongTensor(y_test)
+		y_train = torch.FloatTensor(y_train)
+		y_val = torch.FloatTensor(y_val)
+		y_test = torch.FloatTensor(y_test)
 		
 		# Create TensorDataset
 		train_dataset = torch.utils.data.TensorDataset(X_train, y_train)
